requirements.txt
scrapy
python-slugify
twisted
klein
bs4
tldextract
extruct